Split into per-application files
